backend/requirements.txt
fastapi
uvicorn
pydantic
bcrypt
aiofiles
python-multipart
google-genai